category.js / category.scss 수정
