--- src/Pages/Cart/Components/Cartlist/Cartlist.scss
@@ -1,100 +1,0 @@
-@import '../../Styles/common.scss';
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@1,300&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Nanum+Gothic&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,600;1,300&family=Nanum+Gothic&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,600;1,300&display=swap');
-
-.category{
-  margin-top: 200px;
-  .content{
-    position: relative;
-    width: 1100px;
-    margin: 0 auto;
-  }
-  .goods_pick_list{
-    display: flex;
-    font-size: 14px;
-    padding: 6px 0 10px 10px;
-    .all_count{
-      width: 100%;
-      color: $bm-black;
-    }
-    .pick_list_box{
-      display: flex;
-      justify-content: flex-end;
-      width: 100%;
-      color: $bm-black;
-      font-weight: bold;
-      .pick_list{
-        display: flex;
-        font-family: 'Nanum Gothic', sans-serif;
-        font-size: 13px;
-        .pick_list_sort_recommend{
-          margin: 0 20px;
-          cursor: pointer;
-          color: $bm-beach;
-        }
-       .pick_list_sort{
-         margin: 0 20px;
-         cursor: pointer;
-       }
-      }
-    }
-  }
-}
-
-.item_photo_box{
-  width: 100%;
-}
-.goods_list_item{
-  display: flex;
-  justify-content: center;
-  padding: 15px 0px;
-  width: 100%;
-  line-height: 2;
-  .item_photo_box{
-    overflow: hidden;
-    .img_list{
-      width: 250px;
-      height: 340px;
-      transform:scale(1);
-      transition:.3s;
-      &.img_list:hover{
-      transform:scale(1.2);
-        cursor: pointer;
-      }
-    }
-  }
-  .item_list{
-    display: inline-block;
-    margin: 20px 0px;
-    .item_cont{
-      padding: 0px 10px;
-      .item_info{
-        font-size: 13px;
-        line-height: 1.7;
-      }
-    }
-  }
-}
-
-
-.pagination {
-  display: flex;
-  justify-content: center;
-  margin: 0 auto;
-  padding: 40px 0px;
-  width: 900px;
-  &.pagination ul{
-    display: flex;
-    font-size: 15px;
-    .page_one{
-      font-weight: bold;
-      font-family: 'Nanum Gothic', sans-serif;
-    }
-  }
-  &.pagination li{
-    padding: 0px 15px;
-    font-family: 'Nanum Gothic', sans-serif;
-  }
-}
